requirements.txt
aiohttp==3.9.1
aioprometheus==23.12.0
black==21.10b0
cpm_kernels==1.0.11
#deepspeed==0.3.15+unknown
evaluate==0.4.1
fastapi==0.115.0
file_utils==0.0.1
filelock==3.3.0
fire==0.4.0
gradio==4.44.1
huggingface_hub==0.19.4
impl==0.1.2
jieba==0.42.1
jsonlines==2.0.0
lmdb==1.5.1
modelscope==1.18.1
nltk==3.7
numpy==1.24.4
packaging==24.1
pandas==1.5.3
pathlib2==2.3.7.post1
psutil==5.9.7
pydantic
pytest==8.0.1
pytorch_extension==0.2
pyzmq==23.2.0
ray==2.9.1
regex==2022.10.31
Requests==2.32.3
safetensors==0.3.1
sentence_transformers==3.1.1
sentencepiece==0.1.97
setuptools==65.6.3
Sphinx==8.0.2
strsimpy==0.2.1
tbe==0.0.4
te==1.0.1
tensorboardX==2.6.2.2
tqdm==4.64.1
transformer_engine==1.10.0
transformers==4.36.2
tree_sitter==0.20.1
#triton==2.1.0
uvicorn==0.31.0
wandb==0.13.6
#xformers==0.0.23.post1
